+++ src/Matrix.cpp
@@ -1,7 +1,7 @@
 #include "Matrix.h"
 
 Matrix::Matrix(unsigned int const rowCount, unsigned int const columnCount) :
-        mpData(new float[rowCount * columnCount]), mRowCount(rowCount), mColumnCount(columnCount), mHasAllocation(
+        mpData(new double[rowCount * columnCount]), mRowCount(rowCount), mColumnCount(columnCount), mHasAllocation(
                 true)
 {
 
@@ -9,14 +9,14 @@ Matrix::Matrix(unsigned int const rowCount, unsigned int const columnCount) :
 
 /* explicit */
 Matrix::Matrix(unsigned int const size, unsigned int const rowCount, unsigned int const columnCount) :
-        mpData(new float[size]), mRowCount(rowCount), mColumnCount(columnCount), mHasAllocation(
+        mpData(new double[size]), mRowCount(rowCount), mColumnCount(columnCount), mHasAllocation(
                 true)
 {
 
 }
 
 /* explicit */
-Matrix::Matrix(float* const pData, unsigned int const rowCount, unsigned int const columnCount) :
+Matrix::Matrix(double* const pData, unsigned int const rowCount, unsigned int const columnCount) :
         mpData(pData), mRowCount(rowCount), mColumnCount(columnCount), mHasAllocation(false)
 {
 
@@ -29,13 +29,13 @@ Matrix::~Matrix()
         delete[] mpData;
 }
 
-/* virtual */float&
+/* virtual */double&
 Matrix::at(unsigned int const i, unsigned int const j)
 {
     return mpData[(j * mRowCount) + i];
 }
 
-/* virtual */float const&
+/* virtual */double const&
 Matrix::at(unsigned int const i, unsigned int const j) const
 {
     return mpData[(j * mRowCount) + i];
@@ -53,9 +53,9 @@ Matrix::getRowCount() const
     return mRowCount;
 }
 
-Matrix::operator std::vector<float>() const
+Matrix::operator std::vector<double>() const
 {
-    std::vector<float> elements;
+    std::vector<double> elements;
     elements.reserve(mRowCount * mColumnCount);
 
     for (unsigned int i = 0; i < mColumnCount; ++i)
